Consolidate data flows in .transport.data
- Drop .transport.files submodule, moving contents to .data.
- Update, expand, sort, and copyedit input data docs for consistency.
- Add output data flows to documentation.
