+++ message_ix_models/data/transport/R12/lifetime-ldv.csv
@@ -2,7 +2,6 @@
 #
 # Units: year
 #
-#
 node_loc, year_vtg, value
 R12_AFR,  1990,     25
 R12_CHN,  1990,     15
