+++ message_ix_models/model/transport/key.py
@@ -6,11 +6,12 @@ from genno import Key, Keys, KeySeq
 
 from message_ix_models.report.key import GDP, PRICE_COMMODITY
 
-from . import files
+from .data import iter_files
 
 __all__ = [
     "cg",
     "cost",
+    "exo",
     "fv_cny",
     "fv",
     "gdp_cap",
@@ -127,8 +128,16 @@ t_modes = "t::transport modes"
 #: Model periods.
 y = "y::model"
 
-#: Keys referring to loaded exogenous data flows (data loaded from files).
+#: Keys referring to loaded input data flows (exogenous data loaded from files).
+#: Attributes correspond to the members of :mod:`.transport.data`; see
+#: :doc:`/transport/input` for a complete list.
+#:
+#: .. code-block:: python
+#:
+#:    >>> from message_ix_models.model.transport.key import exo
+#:    >>> exo.act_non_ldv
+#:    <activity:n-t-y:non-ldv+exo>
 exo = Keys()
 
-for name, df in files.iter_files():
+for name, df in iter_files():
     setattr(exo, name, df.key)
